feat(tools): forge-rustc codegen 后端、forge-tests 测试框架（宏族+unicorn）、mini_c Direct/Hir 双后端
- forge-rustc: CodegenBackend 实现（MIR→forge-ir 降级：标量/聚合/enum niche/闭包/dyn vtable/static/drop）+ alloc_runtime 手写 shim + 58 个 e2e 用例
- forge-tests: backend_tests!/encode_golden!/exec! 宏族、feature 矩阵、vendored unicorn 跨架构执行
- forge-object: ELF/PE/Mach-O 写出（COFF addend-4、vtable .data workaround）
- forge-plugin: libloading 动态 ISA 后端加载
- mini_c: Backend::{Direct,Hir} 双后端 + dual_backend_tests 回归 + jit_demo

+++ crates/tools/forge-tests/src/isa/aarch64/disasm.rs
@@ -7,7 +7,7 @@ use code_forge::backend::machine::target::TargetMachine as TargetMachineTrait;
 use code_forge::ir::RegClass;
 
 fn r(n: u32) -> Reg {
-    <Reg as code_forge::ir::PhysReg>::from_index(n as u8, RegClass::Int)
+    <Reg as code_forge::ir::PhysReg>::from_index(n, RegClass::Int)
 }
 
 fn disasm(inst: &Inst) -> String {
